Simplify h.queue by removing local queue
There's no way a local queue setup can work with external worker
processes, so this commit removes support for a local queue. Using
queue-based features locally will now require running nsqd.
