+++ h/test/queue_test.py
@@ -1,12 +1,14 @@
+from pyramid import testing
 from mock import patch
 
-from h.queue.nsq import NSQHelper
+from ..queue import get_reader
+from ..queue import get_writer
 
 
 @patch('gnsq.Reader')
 def test_get_reader_default(fake_reader):
-    qh = NSQHelper({})
-    qh.get_reader('ethics-in-games-journalism', 'channel4')
+    req = testing.DummyRequest()
+    get_reader(req, 'ethics-in-games-journalism', 'channel4')
     fake_reader.assert_called_with('ethics-in-games-journalism',
                                    'channel4',
                                    nsqd_tcp_addresses=['localhost:4150'])
@@ -14,9 +16,11 @@ def test_get_reader_default(fake_reader):
 
 @patch('gnsq.Reader')
 def test_get_reader(fake_reader):
-    qh = NSQHelper({'nsq.reader.addresses': "foo:1234\nbar:4567"})
-    qh.get_reader('ethics-in-games-journalism',
-                  'channel4')
+    req = testing.DummyRequest()
+    req.registry.settings.update({
+        'nsq.reader.addresses': "foo:1234\nbar:4567"
+    })
+    get_reader(req, 'ethics-in-games-journalism', 'channel4')
     fake_reader.assert_called_with('ethics-in-games-journalism',
                                    'channel4',
                                    nsqd_tcp_addresses=['foo:1234',
@@ -25,13 +29,14 @@ def test_get_reader(fake_reader):
 
 @patch('gnsq.Nsqd')
 def test_get_writer_default(fake_nsqd):
-    qh = NSQHelper({})
-    qh.get_writer()
+    req = testing.DummyRequest()
+    get_writer(req)
     fake_nsqd.assert_called_with('localhost', http_port='4151')
 
 
 @patch('gnsq.Nsqd')
 def test_get_writer(fake_nsqd):
-    qh = NSQHelper({'nsq.writer.address': 'philae:2014'})
-    qh.get_writer()
+    req = testing.DummyRequest()
+    req.registry.settings.update({'nsq.writer.address': 'philae:2014'})
+    get_writer(req)
     fake_nsqd.assert_called_with('philae', http_port='2014')
